+++ Modules/ComputerManagementDsc/Examples/Resources/WindowsEventlog/1-SetWindowsEventlogSize.ps1
@@ -1,7 +1,7 @@
 <#
     .EXAMPLE
-    Example script that sets the application eventlog
-    to a maximum size 4096MB.
+    Example script that sets the application Windows Event Log
+    to a maximum size 4096MB, the logmode to 'Circular' and enable it.
 #>
 Configuration Example
 {
@@ -9,12 +9,12 @@ Configuration Example
 
     Node localhost
     {
-        WinEventLog ApplicationEventlogSize
+        WindowsEventLog ApplicationEventlogSize
         {
             LogName            = 'Application'
             IsEnabled          = $true
             LogMode            = 'Circular'
             MaximumSizeInBytes = 4096KB
-        } # End of WinEventLog Resource
+        } # End of Windows Event Log Resource
     } # End of Node
 } # End of Configuration