+++ Modules/ComputerManagementDsc/Examples/Resources/WindowsEventlog/3-EnableWindowsEventLog.ps1
@@ -1,7 +1,7 @@
 <#
     .EXAMPLE
-    Example script that sets the Dsc Analytic eventlog
-    to size maximum size 4096MB, with logmode circular
+    Example script that sets the Dsc Analytic Windows Event Log
+    to size maximum size 4096MB, with logmode 'Retain' and enable it.
 #>
 Configuration Example
 {
@@ -9,13 +9,13 @@ Configuration Example
 
     Node localhost
     {
-        WinEventLog Enable-DscAnalytic
+        WindowsEventLog Enable-DscAnalytic
         {
             LogName             = 'Microsoft-Windows-Dsc/Analytic'
             IsEnabled           = $True
             LogMode             = 'Retain'
             MaximumSizeInBytes  = 4096kb
             LogFilePath         = "%SystemRoot%\System32\Winevt\Logs\Microsoft-Windows-DSC%4Analytic.evtx"
-        } # End of WinEventLog Resource
+        } # End of Windows Event Log Resource
     } # End of Node
 } # End of Configuration